vtt-sync: phase 4 extract token drag and selection to ui/token-interactions.js
Move the token drag lifecycle (prepareTokenDrag, beginTokenDrag,
updateTokenDrag, endTokenDrag, commitDragPreview, applyDragPreview,
syncTokenMeasurement, clearDragCandidate) and marquee selection
(startSelectionBox, updateSelectionBox, finishSelectionBox,
cancelSelectionBox, getTokensInSelectionBox) out of board-interactions.js
into a new ui/token-interactions.js factory. Dependencies (DOM refs,
viewState, selectedTokenIds, rendered placements accessor, boardApi,
and every external helper — getLocalMapPoint, renderTokens,
notifySelectionChanged, measurement callbacks, op/snapshot emitters,
normalizers, etc.) are now passed explicitly instead of read from
the outer closure. board-interactions.js wires the factory up during
mountBoardInteractions and destructures the public methods so existing
pointer-event dispatch sites are unchanged.

dragRenderRafId and dragElements are now closure-scoped inside the
factory; getLocalMapPoint stays in board-interactions.js because it is
used by overlay/template/fog tools elsewhere and is passed in as a dep.

board-interactions.js: 19,498 -> 18,870 lines.
Behavior unchanged. All 261 existing tests still pass. 4 new smoke
tests added in __tests__/token-interactions.test.mjs.
